requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain-groq
python-dotenv
requests